Update README.md
Question 1: Campus Layout

You are asked to determine the layout of a new, small college. The campus will have four structures: an administration structure (A), a bus stop (B), a classroom (C), and a dormitory (D). Each structure (including the bus stop) must be placed somewhere on the grid shown below.

The layout must satisfy the following constraints:

i. The bus stop (B) must be adjacent to the road.

    The administration structure (A) and the classroom (C) must both be adjacent to the bus stop (B).

    The classroom (C) must be adjacent to the dormitory (D).

    The administration structure (A) must not be adjacent to the dormitory (D).

    The administration structure (A) must not be on a hill.

    The dormitory (D) must be on a hill or adjacent to the road.

    All structures must be in different grid squares.

Here, adjacent means that the structures must share a grid edge, not just a corner.

We recommend you work out the solutions to the following questions on a sheet of scratch paper, and then enter your results below.